move const to separate file
