+++ nest-city-account/src/admin/utils/constants.ts
@@ -1,0 +1,1 @@
+export const cognitoSyncConfigDbkey = 'SYNC_COGNITO_TO_DB'
